Remove the warning
...which would appear in the eventuality that the Twitter team changes
the structure of the page and TLD can't detect the lists with tweets
and replies any more, to tell the user to contact me, in case I didn't
know already.
I added the warning in one of the early versions, when TLD only
cleaned the links from tweets. Since then, TLD has grown to do more
than that and it will continue to do so. I tried to improve the
warning to keep up with the TLD updates but it keeps getting triggered
unexpectedly and it has become a nuissance for me to fix it. Because
of that and the fact that it is slowly becoming obsolete, I decided to
eliminate it and automate my browser to check the page structure
regularly.